backend/requirements.txt
# Knowledge AI - Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Vector database
chromadb>=0.4.22

# HTTP client
httpx>=0.26.0

# File system watching
watchdog>=3.0.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pdfplumber>=0.10.3

# Security
cryptography>=41.0.0

# Type checking
pydantic>=2.5.0

# Utilities
python-multipart>=0.0.6
